refactory gateway trigger

+++ store-service/src/main/java/com/kii/beehive/business/ruleengine/TriggerManager.java
@@ -4,9 +4,11 @@ import javax.annotation.PostConstruct;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,18 +23,30 @@ import com.kii.beehive.business.helper.TriggerCreator;
 import com.kii.beehive.business.manager.AppInfoManager;
 import com.kii.beehive.business.manager.ThingTagManager;
 import com.kii.beehive.business.service.ThingIFInAppService;
+import com.kii.beehive.portal.common.utils.ThingIDTools;
 import com.kii.beehive.portal.event.EventListener;
 import com.kii.beehive.portal.exception.EntryNotFoundException;
 import com.kii.beehive.portal.jdbc.dao.GlobalThingSpringDao;
+import com.kii.beehive.portal.jdbc.entity.GlobalThingInfo;
 import com.kii.beehive.portal.service.EventListenerDao;
 import com.kii.beehive.portal.service.LocalTriggerRecordDao;
 import com.kii.extension.ruleengine.EngineService;
 import com.kii.extension.ruleengine.drools.entity.ThingStatusInRule;
 import com.kii.extension.ruleengine.schedule.ScheduleService;
 import com.kii.extension.ruleengine.service.TriggerRecordDao;
+import com.kii.extension.ruleengine.store.trigger.BeehiveTriggerType;
+import com.kii.extension.ruleengine.store.trigger.CommandToThing;
+import com.kii.extension.ruleengine.store.trigger.CommandToThingInGW;
+import com.kii.extension.ruleengine.store.trigger.ExecuteTarget;
+import com.kii.extension.ruleengine.store.trigger.GatewayTriggerRecord;
+import com.kii.extension.ruleengine.store.trigger.GroupTriggerRecord;
 import com.kii.extension.ruleengine.store.trigger.SimpleTriggerRecord;
 import com.kii.extension.ruleengine.store.trigger.TriggerRecord;
+import com.kii.extension.ruleengine.store.trigger.WhenType;
+import com.kii.extension.ruleengine.store.trigger.condition.LogicCol;
 import com.kii.extension.sdk.entity.thingif.Action;
+import com.kii.extension.sdk.entity.thingif.EndNodeOfGateway;
+import com.kii.extension.sdk.entity.thingif.ThingCommand;
 import com.kii.extension.sdk.entity.thingif.ThingOfKiiCloud;
 import com.kii.extension.sdk.entity.thingif.ThingStatus;
 
@@ -46,9 +60,6 @@ public class TriggerManager {
 
 	@Autowired
 	private AppInfoManager appInfoManager;
-
-	@Autowired
-	private LocalTriggerManager localTriggerManager;
 
 	@Autowired
 	private TriggerRecordDao triggerDao;
@@ -92,11 +103,16 @@ public class TriggerManager {
 	@PostConstruct
 	public void init() {
 		List<TriggerRecord> recordList = triggerDao.getAllTrigger();
+
+
 		scheduleService.startSchedule();
 
 		recordList.forEach(record -> {
 
 			try {
+				if(record.getType()== BeehiveTriggerType.Gateway){
+					return;
+				}
 				creator.addTriggerToEngine(record);
 			} catch (Exception e) {
 				e.printStackTrace();
@@ -130,35 +146,117 @@ public class TriggerManager {
 	}
 
 
-	public String createTrigger(TriggerRecord record) {
+	public TriggerRecord createTrigger(TriggerRecord record) {
 
-		ThingOfKiiCloud gatewayOfKiiCloud = null;
-		try {
-			gatewayOfKiiCloud = localTriggerManager.checkLocalRule(record);
-		} catch (Exception e) {
-			e.printStackTrace();
-			log.error("checkLocalRule error ", e);
+
+		if(checkLocalRule(record)){
+			try {
+				return createGatewayRecord((GroupTriggerRecord) record);
+			}catch(IllegalStateException e){
+				log.warn("invalid gateway trigger param");
+			}
 		}
 
-		if( ! StringUtils.isEmpty(record.getId()) ){// saved local rule
-			//action
-			Map<String, Action> actions = new HashMap<>();
-			Action action = new Action();
-			actions.put("createTrigger", action);
-			action.setField("triggerJson", record);
+		triggerDao.addKiiEntity(record);
 
-			localTriggerManager.sendGatewayCommand(gatewayOfKiiCloud.getFullKiiThingID(), actions);
+		creator.addTriggerToEngine(record);
+		return record;
 
-		}else {
-
-			triggerDao.addKiiEntity(record);
-
-			creator.addTriggerToEngine(record);
-		}
-
-		return record.getId();
 
 	}
+
+
+	private boolean checkLocalRule(TriggerRecord record) {
+
+		if( ! (record instanceof GroupTriggerRecord
+				&& record.getPredicate().getTriggersWhen().equals(WhenType.CONDITION_TRUE)
+				&& record.getPredicate().getSchedule() == null
+				&& ! ( record.getPredicate().getCondition()  instanceof LogicCol) ) ) {
+
+			return false;
+		}
+
+		GroupTriggerRecord  groupRecord=(GroupTriggerRecord)record;
+
+		//source only one thing
+		if( ( groupRecord.getSource().getTagList() != null && groupRecord.getSource().getTagList().size() > 0 )
+				||  ( groupRecord.getSource().getThingList() != null && groupRecord.getSource().getThingList().size() > 1 )){
+			return false;
+		}
+
+
+
+		return true ;
+	}
+
+
+	private GatewayTriggerRecord createGatewayRecord(GroupTriggerRecord  groupRecord){
+
+
+		GatewayTriggerRecord triggerRecord = new GatewayTriggerRecord();
+		triggerRecord.setPolicy(groupRecord.getPolicy());
+		triggerRecord.setRecordStatus(groupRecord.getRecordStatus());
+		triggerRecord.setPreparedCondition(groupRecord.getPreparedCondition());
+		triggerRecord.setPredicate(groupRecord.getPredicate());
+		triggerRecord.setTargetParamList(groupRecord.getTargetParamList());
+
+		triggerRecord.setUserID(groupRecord.getUserID());
+		triggerRecord.setDescription(groupRecord.getDescription());
+
+		GlobalThingInfo sourceThing = globalThingDao.findByID(groupRecord.getSource().getThingList().get(0));
+		triggerRecord.getSource().getVendorThingIdList().add(sourceThing.getVendorThingID());
+
+		//
+		ThingOfKiiCloud gatewayOfKiiCloud = thingIFService.getThingGateway(sourceThing.getFullKiiThingID());
+		String thingID=gatewayOfKiiCloud.getThingID();
+
+//		String thingID="th.f83120e36100-a269-5e11-bf4b-0c5b4813";
+		String venderThingID=globalThingDao.getThingByFullKiiThingID(sourceThing.getKiiAppID(), thingID).getVendorThingID();
+
+		String fullKiiThingID=ThingIDTools.joinFullKiiThingID(sourceThing.getKiiAppID(), thingID);
+
+		List<EndNodeOfGateway> allEndNodesOfGateway = thingIFService.getAllEndNodesOfGateway(fullKiiThingID);
+		Map<String, EndNodeOfGateway> allEndNodesOfGatewayMap = new HashMap<>();
+		allEndNodesOfGateway.forEach(endNodeOfGateway -> allEndNodesOfGatewayMap.put(endNodeOfGateway.getVendorThingID(), endNodeOfGateway));
+
+		List<ExecuteTarget> targets = groupRecord.getTargets();
+
+//		targets:
+		for(ExecuteTarget target:targets){
+			switch (target.getType()) {
+
+				case "ThingCommand":
+					CommandToThing command=(CommandToThing)target;
+					CommandToThingInGW cmdInGW=new CommandToThingInGW();
+					cmdInGW.setCommand(command.getCommand());
+					cmdInGW.getSelector().setVendorThingIdList(new ArrayList<>());
+					Set<GlobalThingInfo> thingList = thingTagService.getThingInfos(command.getSelector());
+
+					for (GlobalThingInfo thing : thingList){
+						if( allEndNodesOfGatewayMap.get(thing.getVendorThingID()) == null ){
+							throw new IllegalStateException();
+						}
+						cmdInGW.getSelector().getVendorThingIdList().add(thing.getVendorThingID());
+					}
+					triggerRecord.addTarget(cmdInGW);
+
+					break;
+				case "HttpApiCall":
+					triggerRecord.addTarget(target);
+					break;
+			}
+		}
+		triggerRecord.setGatewayVendorThingID(venderThingID);
+		triggerRecord.setGatewayFullKiiThingID(fullKiiThingID);
+		triggerDao.addKiiEntity(triggerRecord);
+
+		sendGatewayCommand(triggerRecord,GatewayCommand.createTrigger);
+
+		return triggerRecord;
+
+	}
+
+
 
 
 	public Map<String, Object> getRuleEngingDump() {
@@ -173,17 +271,36 @@ public class TriggerManager {
 
 
 	public void disableTrigger(String triggerID) {
+		TriggerRecord  record= triggerDao.getTriggerRecord(triggerID);
+
 		triggerDao.disableTrigger(triggerID);
 
-		service.disableTrigger(triggerID);
+		if(record.getType()==BeehiveTriggerType.Gateway){
 
+			sendGatewayCommand((GatewayTriggerRecord) record, GatewayCommand.disableTrigger);
+
+		}else {
+
+			service.disableTrigger(triggerID);
+		}
 	}
 
 
 	public void enableTrigger(String triggerID) {
+
+
+		TriggerRecord  record= triggerDao.getTriggerRecord(triggerID);
+
 		triggerDao.enableTrigger(triggerID);
 
-		service.enableTrigger(triggerID);
+		if(record.getType()==BeehiveTriggerType.Gateway){
+
+			sendGatewayCommand((GatewayTriggerRecord) record, GatewayCommand.enableTrigger);
+
+		}else {
+
+			service.enableTrigger(triggerID);
+		}
 	}
 
 	public List<TriggerRecord> getTriggerListByUserId(Long userId) {
@@ -229,17 +346,59 @@ public class TriggerManager {
 
 	public void deleteTrigger(String triggerID) {
 
+
+		TriggerRecord  record= triggerDao.getTriggerRecord(triggerID);
+
+		if(record.getType()==BeehiveTriggerType.Gateway){
+
+
+			sendGatewayCommand((GatewayTriggerRecord) record, GatewayCommand.deleteTrigger);
+
+		}else {
+			service.removeTrigger(triggerID);
+
+			scheduleService.removeManagerTaskForSchedule(triggerID);
+
+			List<EventListener> eventListenerList = eventListenerDao.getEventListenerByTargetKey(triggerID);
+			for (EventListener eventListener : eventListenerList) {
+				eventListenerDao.removeEntity(eventListener.getId());
+			}
+		}
+
 		triggerDao.deleteTriggerRecord(triggerID);
 
-		service.removeTrigger(triggerID);
-
-		scheduleService.removeManagerTaskForSchedule(triggerID);
-
-		List<EventListener> eventListenerList = eventListenerDao.getEventListenerByTargetKey(triggerID);
-		for (EventListener eventListener : eventListenerList) {
-			eventListenerDao.removeEntity(eventListener.getId());
-		}
 	}
 
+	private enum GatewayCommand{
+
+		deleteTrigger,disableTrigger,enableTrigger,createTrigger;
+	}
+
+
+	private  void sendGatewayCommand(GatewayTriggerRecord  record,GatewayCommand act ) {
+
+
+		String triggerID=record.getTriggerID();
+
+		String fullThingID=record.getGatewayFullKiiThingID();
+
+		Map<String, Action> actions = new HashMap<>();
+		Action action = new Action();
+		actions.put(act.name(), action);
+		action.setField("triggerID", triggerID);
+
+		//command								send to gateway
+		ThingCommand command = new ThingCommand();
+		command.setSchema("gateway");
+		command.setTitle("trigger");
+		//action
+		command.setActions(Arrays.asList(actions));
+
+		ThingIDTools.ThingIDCombine combine = ThingIDTools.splitFullKiiThingID(fullThingID);
+
+		command.setUserID(appInfoManager.getDefaultOwer(combine.kiiAppID).getUserID());
+		thingIFService.sendCommand(command, fullThingID);
+
+	}
 
 }
